--- src/main/java/kr/hhplus/be/server/api/concert/dto/ConcertScheduleResponse.java
@@ -1,13 +1,0 @@
-package kr.hhplus.be.server.api.concert.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDate;
-
-@Getter
-@Builder
-public class ConcertScheduleResponse {
-    private Long scheduleId;
-    private LocalDate concertDate;
-}
