+++ src/main/java/kr/hhplus/be/server/concert/domain/soldoutRanking/ConcertSoldoutRankingRepository.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.concert.domain.soldoutRanking;
+
+import java.util.List;
+
+public interface ConcertSoldoutRankingRepository {
+    List<String> getSoldoutRanking(String yearMonth, int topN);
+    void addSoldoutRanking(String yearMonth, Long scheduleId, double elapsedSeconds);
+    boolean isAlreadyRanked(String yearMonth, Long scheduleId);
+    void evictSoldoutRanking(String yearMonth);
+}
