+++ src/main/java/kr/hhplus/be/server/concert/dto/response/ConcertScheduleResponse.java
@@ -1,0 +1,13 @@
+package kr.hhplus.be.server.concert.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.time.LocalDate;
+
+@Getter
+@Builder
+public class ConcertScheduleResponse {
+    private Long scheduleId;
+    private LocalDate concertDate;
+}
