--- src/main/java/kr/hhplus/be/server/domain/concert/event/ConcertSoldoutEvent.java
@@ -1,12 +1,0 @@
-package kr.hhplus.be.server.domain.concert.event;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class ConcertSoldoutEvent {
-
-    private final Long scheduleId;
-
-}
