--- src/main/java/kr/hhplus/be/server/infrastructure/persistence/concert/ConcertScheduleEntity.java
@@ -1,36 +1,0 @@
-package kr.hhplus.be.server.infrastructure.persistence.concert;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-
-@Entity
-@Table(name = "concert_schedules")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class ConcertScheduleEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "schedule_id")
-    private Long scheduleId;
-
-    @Column(name = "concert_date", nullable = false)
-    private LocalDate concertDate;
-
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-}
