--- src/main/java/kr/hhplus/be/server/infrastructure/persistence/concert/SeatEntity.java
@@ -1,46 +1,0 @@
-package kr.hhplus.be.server.infrastructure.persistence.concert;
-
-import jakarta.persistence.*;
-import kr.hhplus.be.server.domain.concert.SeatStatus;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "seats")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class SeatEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "seat_id")
-    private Long seatId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "schedule_id", nullable = false)
-    private ConcertScheduleEntity concertSchedule;
-
-    @Column(name = "seat_number", nullable = false)
-    private int seatNumber;
-
-    @Column(name = "price", nullable = false)
-    private int price;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "status", nullable = false)
-    private SeatStatus status;
-
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-}
